+++ providers/common/digests/digest_common.c
@@ -8,7 +8,7 @@
  */
 
 #include "openssl/err.h"
-#include "internal/digestcommon.h"
+#include "prov/digestcommon.h"
 #include "internal/providercommonerr.h"
 
 int digest_default_get_params(OSSL_PARAM params[], size_t blksz, size_t paramsz,
