+++ providers/common/include/prov/digestcommon.h
