--- providers/default/digests/build.info
@@ -1,16 +1,0 @@
-$GOAL=../../libimplementations.a
-
-IF[{- !$disabled{blake2} -}]
-  SOURCE[$GOAL]=\
-          blake2_prov.c blake2b_prov.c blake2s_prov.c
-ENDIF
-
-IF[{- !$disabled{sm3} -}]
-  SOURCE[$GOAL]=\
-          sm3_prov.c
-ENDIF
-
-IF[{- !$disabled{md5} -}]
-  SOURCE[$GOAL]=\
-          md5_prov.c md5_sha1_prov.c
-ENDIF
