+++ providers/implementations/digests/blake2_prov.c
@@ -8,8 +8,8 @@
  */
 
 #include <openssl/crypto.h>
-#include "internal/blake2.h"
-#include "internal/digestcommon.h"
+#include "prov/blake2.h"
+#include "prov/digestcommon.h"
 #include "internal/provider_algs.h"
 
 OSSL_OP_digest_init_fn blake2s256_init;
