+++ providers/implementations/digests/wp_prov.c
@@ -8,11 +8,11 @@
  */
 
 #include <openssl/crypto.h>
-#include <openssl/ripemd.h>
-#include "internal/digestcommon.h"
+#include <openssl/whrlpool.h>
+#include "prov/digestcommon.h"
 #include "internal/provider_algs.h"
 
-/* ripemd160_functions */
-IMPLEMENT_digest_functions(ripemd160, RIPEMD160_CTX,
-                           RIPEMD160_CBLOCK, RIPEMD160_DIGEST_LENGTH, 0,
-                           RIPEMD160_Init, RIPEMD160_Update, RIPEMD160_Final)
+/* wp_functions */
+IMPLEMENT_digest_functions(wp, WHIRLPOOL_CTX,
+                           WHIRLPOOL_BBLOCK / 8, WHIRLPOOL_DIGEST_LENGTH, 0,
+                           WHIRLPOOL_Init, WHIRLPOOL_Update, WHIRLPOOL_Final)
