--- providers/legacy/digests/build.info
@@ -1,21 +1,0 @@
-$GOAL=../../liblegacy.a
-
-IF[{- !$disabled{md2} -}]
-  SOURCE[$GOAL]=md2_prov.c
-ENDIF
-
-IF[{- !$disabled{md4} -}]
-  SOURCE[$GOAL]=md4_prov.c
-ENDIF
-
-IF[{- !$disabled{mdc2} -}]
-  SOURCE[$GOAL]=mdc2_prov.c
-ENDIF
-
-IF[{- !$disabled{whirlpool} -}]
-  SOURCE[$GOAL]=wp_prov.c
-ENDIF
-
-IF[{- !$disabled{rmd160} -}]
-  SOURCE[$GOAL]=ripemd_prov.c
-ENDIF
